Improve speed of query finding Albums a Track has been on.
Including scrobble_counts.

For #164
